+++ morpha/morpha/models/__init__.py
@@ -2,6 +2,8 @@
 
 from .irecord import IRecord
 from .team_preview_record import TeamPreviewRecord
+from .battle_log import BattleLog
 
-__all__ = ['IRecord',
+__all__ = ['BattleLog',
+           'IRecord',
            'TeamPreviewRecord']