+++ morpha/morpha/parsers/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .irecord_parser import IRecordParser
-from .battle_log_parser import BattleLogParser
-
-__all__ = ['BattleLogParser',
-           'IRecordParser']