+++ morpha/documentation/release_notes.md
@@ -2,5 +2,4 @@
 
 ## v0.1.0 (Pending)
 ##### Features
-- Added IRecordParser, TeamPreviewRecordParser, and BattleLogParser (`ME-630`).
-- Added IRecord and TeamPreviewRecord models (`ME-630`).
+- Added BattleLog, IRecord, and TeamPreviewRecord models (`ME-630`).